improve code quality

+++ server/util.c
@@ -2,12 +2,16 @@
 // Created by 朱逸尘 on 2017/8/4.
 //
 
-#include "util.h"
 #include <string.h>
-#include <sys/types.h>
 #include <unistd.h>
 #include <errno.h>
 
+/**
+ * write buffer to target fd
+ * @param conn_fd the fd that write to
+ * @param buffer  char buffer
+ * @param length  the length of buffer
+ */
 void non_blocking_write(int conn_fd, char *buffer, int length) {
     int has_write = 0;
     int once_write = 0;
@@ -29,6 +33,10 @@ void non_blocking_write(int conn_fd, char *buffer, int length) {
     }
 }
 
+/**
+ * remove char' ' in the head and tail of the target str
+ * @param str the target str
+ */
 void trim(char *str) {
     int i = 0;
     while (str[i] != '\0') {
@@ -46,8 +54,15 @@ void trim(char *str) {
     }
 }
 
+/**
+ * judge if the target str is end with ch
+ *
+ * @param str the target str
+ * @param ch  the check char
+ * @return 1 is false, 0 is true
+ */
 int is_end_with_ch(char* str, char ch) {
-    int length = strlen(str);
+    int length = (int)strlen(str);
     if (str[length-1] == ch) {
         return 1;
     }
